requirements.txt
Django==5.0.6
djangorestframework==3.15.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pillow==10.3.0
djangorestframework-simplejwt==5.3.1
django-filter==24.2
drf-yasg==1.21.7
django-cors-headers==4.3.1
stripe==9.10.0
forex-python==1.8
requests==2.32.3
coverage==7.5.3
